Legacy_AEFI_Acquisition/EFImagingBench_App/tests/archives/requirements_analysis.txt
# Dépendances pour l'analyse spectrale de l'accéléromètre LSM9D
# =================================================================

# Calcul scientifique et traitement du signal
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualisation
matplotlib>=3.5.0
seaborn>=0.11.0

# Interface série (pour le script d'acquisition)
pyserial>=3.5

# Gestion des fichiers et chemins
pathlib2>=2.3.6; python_version<'3.4'

# Utilitaires
argparse>=1.4.0; python_version<'2.7'

# Optionnel : pour de meilleures performances
# numba>=0.56.0  # Accélération JIT pour les calculs intensifs
# plotly>=5.0.0  # Graphiques interactifs alternatifs 